@@ -1,35 +1,38 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [project]
 name = "pdf-svc"
-version = "0.1.0"
-description = "Event-driven pdf generator service (FastStream + Redis + NATS)"
+version = "1.0.0"
+description = "Event-driven PDF generation microservice"
 readme = "README.md"
-requires-python = ">=3.11"
-license = { text = "MIT" }
-authors = [{ name = "Inteligente", email = "dev@example.com" }]
-keywords = ["pdf", "faststream", "nats", "redis", "microservice", "event-driven"]
-classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Developers",
-    "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Topic :: Software Development :: Libraries :: Python Modules",
+requires-python = ">=3.12"
+license = "MIT"
+authors = [
+    { name = "Inteligente" }
 ]
 
 dependencies = [
-    "faststream[nats,redis]>=0.5.0",
-    "pydantic>=2.5.0",
-    "pydantic-settings>=2.1.0",
-    "structlog>=24.1.0",
-    "colorama>=0.4.6",
+    # Async & messaging
+    "nats-py>=2.9.0",
+    "redis>=5.0.0",
+    # PDF processing
     "pymupdf>=1.24.0",
     "segno>=1.6.0",
-    "pillow>=10.2.0",
+    "pillow>=10.0.0",
+    # Data validation
+    "pydantic>=2.5.0",
+    "pydantic-settings>=2.1.0",
+    # HTTP client
     "httpx>=0.27.0",
-    "python-dotenv>=1.0.0",
+    # Logging
+    "structlog>=24.1.0",
+    "colorama>=0.4.6",
+    # Serialization
     "orjson>=3.9.0",
-    "tenacity>=8.2.0",
-    "aiocache>=0.12.0",
+    # Environment
+    "python-dotenv>=1.0.0",
 ]
 
 [project.optional-dependencies]
@@ -37,65 +40,47 @@ dev = [
     "pytest>=8.0.0",
     "pytest-asyncio>=0.23.0",
     "pytest-cov>=4.1.0",
-    "pytest-mock>=3.12.0",
-    "ruff>=0.2.0",
-    "mypy>=1.8.0",
-    "pre-commit>=3.6.0",
     "faker>=22.0.0",
+    "ruff>=0.1.0",
+    "mypy>=1.8.0",
 ]
 
 [project.scripts]
 pdf-svc = "pdf_svc.main:main"
 
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
 [tool.hatch.build.targets.wheel]
 packages = ["src/pdf_svc"]
 
 [tool.pytest.ini_options]
+testpaths = ["tests"]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"
-testpaths = ["tests"]
-python_files = ["test_*.py", "*_test.py"]
-python_classes = ["Test*"]
-python_functions = ["test_*"]
-addopts = ["-v", "--tb=short", "-ra", "--strict-markers"]
 markers = [
     "unit: Unit tests",
     "integration: Integration tests",
-    "slow: Slow running tests",
-]
-filterwarnings = [
-    "ignore::DeprecationWarning",
-    "ignore::pytest.PytestUnraisableExceptionWarning",
 ]
 
 [tool.ruff]
-target-version = "py311"
 line-length = 100
+target-version = "py312"
 src = ["src", "tests"]
 
 [tool.ruff.lint]
-select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
-ignore = ["E501", "B008", "B904"]
-
-[tool.ruff.lint.isort]
-known-first-party = ["pdf_svc"]
+select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
+ignore = ["E501"]
 
 [tool.mypy]
-python_version = "3.11"
-warn_return_any = true
-warn_unused_configs = true
-disallow_untyped_defs = true
-plugins = ["pydantic.mypy"]
-
-[[tool.mypy.overrides]]
-module = ["fitz.*", "segno.*"]
+python_version = "3.12"
+strict = true
 ignore_missing_imports = true
 
 [tool.coverage.run]
 source = ["src/pdf_svc"]
-branch = true
-omit = ["*/tests/*", "*/__init__.py"]
+omit = ["tests/*"]
+
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "if TYPE_CHECKING:",
+    "raise NotImplementedError",
+]
