@@ -1,191 +1,209 @@
 """
-Event publishers for NATS.
+NATS Event Publishers for pdf-svc.
 """
 
 from __future__ import annotations
 
-from datetime import datetime, timezone
-from typing import Any, Optional
+from typing import TYPE_CHECKING
 from uuid import UUID
 
+import structlog
 from nats.aio.client import Client as NatsClient
 
-from pdf_svc.config.settings import get_settings
+from pdf_svc.config.settings import Settings
 from pdf_svc.models.events import (
-    PdfJobStatusPayload,
-    PdfJobStatusResponse,
-    PdfProcessCompleted,
-    PdfProcessCompletedPayload,
-    PdfProcessFailed,
-    PdfProcessFailedPayload,
+    PdfBatchCompleted,
+    PdfBatchFailed,
+    PdfBatchResultPayload,
+    PdfItemCompleted,
+    PdfItemCompletedPayload,
+    PdfItemFailed,
+    PdfItemFailedPayload,
+    PdfItemResult,
 )
-from pdf_svc.models.job import Job
-from pdf_svc.shared.logger import get_logger
+from pdf_svc.models.job import BatchJob
 
-logger = get_logger(__name__)
+if TYPE_CHECKING:
+    pass
+
+logger = structlog.get_logger()
 
 
 class PdfEventPublisher:
-    """
-    Publisher for PDF service events.
-    """
+    """Publisher for pdf-svc events."""
 
     def __init__(
         self,
         nats_client: NatsClient,
-        completed_subject: str = "pdf.process.completed",
-        failed_subject: str = "pdf.process.failed",
-        status_subject: str = "pdf.job.status.response",
-    ):
-        self._nats = nats_client
-        self._completed_subject = completed_subject
-        self._failed_subject = failed_subject
-        self._status_subject = status_subject
-
-    async def publish_completed(
-        self,
-        job: Job,
-        file_id: UUID,
-        file_name: str,
-        file_size: int,
-        file_hash: Optional[str] = None,
-        download_url: Optional[str] = None,
-        processing_time_ms: int = 0,
+        settings: Settings,
     ) -> None:
         """
-        Publish PDF processing completed event.
+        Initialize event publisher.
 
         Args:
-            job: Completed job
-            file_id: Uploaded file ID
-            file_name: File name
-            file_size: File size in bytes
-            file_hash: File hash
-            download_url: Download URL
-            processing_time_ms: Processing time in milliseconds
+            nats_client: Connected NATS client
+            settings: Application settings
         """
-        event = PdfProcessCompleted(
-            payload=PdfProcessCompletedPayload(
-                pdf_job_id=job.id,
-                file_id=file_id,
-                file_name=file_name,
-                file_hash=file_hash,
-                file_size_bytes=file_size,
-                download_url=download_url,
-                created_at=datetime.now(timezone.utc),
-                processing_time_ms=processing_time_ms,
+        self.nats = nats_client
+        self.settings = settings
+
+    async def publish_batch_completed(self, job: BatchJob) -> None:
+        """
+        Publish batch processing completed event.
+
+        Args:
+            job: Completed BatchJob
+        """
+        log = logger.bind(job_id=str(job.job_id))
+
+        # Build item results
+        items = [
+            PdfItemResult(
+                item_id=item.item_id,
+                user_id=item.user_id,
+                serial_code=item.serial_code,
+                status=item.status.value,
+                data=item.data.model_dump(mode="json") if item.data else None,
+                error=item.error.model_dump(mode="json") if item.error else None,
+            )
+            for item in job.items
+        ]
+
+        event = PdfBatchCompleted(
+            payload=PdfBatchResultPayload(
+                job_id=job.job_id,
+                status=job.status.value,
+                total_items=job.total_items,
+                success_count=job.success_count,
+                failed_count=job.failed_count,
+                items=items,
+                processing_time_ms=job.processing_time_ms,
             )
         )
 
-        data = event.model_dump_json().encode()
-        await self._nats.publish(self._completed_subject, data)
-
-        logger.info(
-            "event_published",
-            event_type="pdf.process.completed",
-            job_id=str(job.id),
-            file_id=str(file_id),
+        await self.nats.publish(
+            self.settings.pdf_svc.completed_subject,
+            event.model_dump_json().encode(),
         )
 
-    async def publish_failed(
+        log.info(
+            "batch_completed_published",
+            subject=self.settings.pdf_svc.completed_subject,
+            success_count=job.success_count,
+            failed_count=job.failed_count,
+        )
+
+    async def publish_batch_failed(
         self,
-        job: Job,
-        stage: str,
+        job_id: UUID,
         error: str,
-        error_code: Optional[str] = None,
-        details: Optional[dict[str, Any]] = None,
+        code: str | None = None,
     ) -> None:
         """
-        Publish PDF processing failed event.
+        Publish batch processing failed event (entire batch failed).
 
         Args:
-            job: Failed job
-            stage: Stage where failure occurred
+            job_id: Job UUID
             error: Error message
-            error_code: Error code
-            details: Additional error details
+            code: Optional error code
         """
-        event = PdfProcessFailed(
-            payload=PdfProcessFailedPayload(
-                pdf_job_id=job.id,
-                stage=stage,
-                error=error,
-                error_code=error_code,
-                details=details,
-            )
+        log = logger.bind(job_id=str(job_id))
+
+        event = PdfBatchFailed(
+            payload={
+                "job_id": str(job_id),
+                "error": error,
+                "code": code,
+            }
         )
 
-        data = event.model_dump_json().encode()
-        await self._nats.publish(self._failed_subject, data)
+        await self.nats.publish(
+            self.settings.pdf_svc.failed_subject,
+            event.model_dump_json().encode(),
+        )
 
-        logger.info(
-            "event_published",
-            event_type="pdf.process.failed",
-            job_id=str(job.id),
-            stage=stage,
+        log.error(
+            "batch_failed_published",
+            subject=self.settings.pdf_svc.failed_subject,
             error=error,
         )
 
-    async def publish_status(self, job: Job) -> None:
+    async def publish_item_completed(
+        self,
+        job_id: UUID,
+        item_id: UUID,
+        user_id: UUID,
+        serial_code: str,
+        data: dict,
+    ) -> None:
         """
-        Publish job status response.
+        Publish individual item completed event (for real-time tracking).
 
         Args:
-            job: Job to report status for
+            job_id: Parent job UUID
+            item_id: Item UUID
+            user_id: User UUID
+            serial_code: Item serial code
+            data: Item result data
         """
-        result_payload = None
-        error_payload = None
-
-        if job.result:
-            result_payload = PdfProcessCompletedPayload(
-                pdf_job_id=job.id,
-                file_id=job.result.file_id,
-                file_name=job.result.file_name,
-                file_hash=job.result.file_hash,
-                file_size_bytes=job.result.file_size_bytes,
-                download_url=job.result.download_url,
-                created_at=job.result.created_at,
-                processing_time_ms=int((job.duration_seconds or 0) * 1000),
-            )
-
-        if job.error:
-            error_payload = PdfProcessFailedPayload(
-                pdf_job_id=job.id,
-                stage=job.error.stage.value,
-                error=job.error.message,
-                details=job.error.details,
-            )
-
-        event = PdfJobStatusResponse(
-            payload=PdfJobStatusPayload(
-                pdf_job_id=job.id,
-                status=job.status.value,
-                stage=job.stage.value if job.stage else None,
-                progress_pct=job.progress_pct,
-                created_at=job.created_at,
-                updated_at=job.updated_at,
-                completed_at=job.completed_at,
-                result=result_payload,
-                error=error_payload,
+        event = PdfItemCompleted(
+            payload=PdfItemCompletedPayload(
+                job_id=job_id,
+                item_id=item_id,
+                user_id=user_id,
+                serial_code=serial_code,
+                data=data,
             )
         )
 
-        data = event.model_dump_json().encode()
-        await self._nats.publish(self._status_subject, data)
+        await self.nats.publish(
+            self.settings.pdf_svc.item_completed_subject,
+            event.model_dump_json().encode(),
+        )
 
         logger.debug(
-            "status_published",
-            job_id=str(job.id),
-            status=job.status.value,
-            progress=job.progress_pct,
+            "item_completed_published",
+            job_id=str(job_id),
+            item_id=str(item_id),
+            serial_code=serial_code,
         )
 
+    async def publish_item_failed(
+        self,
+        job_id: UUID,
+        item_id: UUID,
+        user_id: UUID,
+        serial_code: str,
+        error: dict,
+    ) -> None:
+        """
+        Publish individual item failed event.
 
-def create_pdf_publisher(nats_client: NatsClient) -> PdfEventPublisher:
-    """Factory function to create PDF event publisher."""
-    settings = get_settings()
-    return PdfEventPublisher(
-        nats_client=nats_client,
-        completed_subject=settings.pdf_svc.completed_subject,
-        failed_subject=settings.pdf_svc.failed_subject,
-    )
+        Args:
+            job_id: Parent job UUID
+            item_id: Item UUID
+            user_id: User UUID
+            serial_code: Item serial code
+            error: Error information
+        """
+        event = PdfItemFailed(
+            payload=PdfItemFailedPayload(
+                job_id=job_id,
+                item_id=item_id,
+                user_id=user_id,
+                serial_code=serial_code,
+                error=error,
+            )
+        )
+
+        await self.nats.publish(
+            self.settings.pdf_svc.item_failed_subject,
+            event.model_dump_json().encode(),
+        )
+
+        logger.debug(
+            "item_failed_published",
+            job_id=str(job_id),
+            item_id=str(item_id),
+            serial_code=serial_code,
+        )
