@@ -1,96 +1,94 @@
 """
-Job repository for Redis persistence.
+Redis Job Repository - Persistence for batch jobs.
 """
 
 from __future__ import annotations
 
-from typing import Optional, Protocol
+from typing import TYPE_CHECKING
 from uuid import UUID
 
 import orjson
+import structlog
 from redis.asyncio import Redis
 
-from pdf_svc.config.settings import get_settings
-from pdf_svc.models.job import Job
-from pdf_svc.shared.logger import get_logger
+from pdf_svc.config.settings import Settings
+from pdf_svc.models.job import BatchJob, JobStatus
 
-logger = get_logger(__name__)
+if TYPE_CHECKING:
+    pass
 
-
-class JobRepositoryProtocol(Protocol):
-    """Protocol for job repository implementations."""
-
-    async def save(self, job: Job) -> None:
-        """Save job to storage."""
-        ...
-
-    async def get(self, job_id: UUID) -> Optional[Job]:
-        """Get job by ID."""
-        ...
-
-    async def delete(self, job_id: UUID) -> bool:
-        """Delete job by ID."""
-        ...
-
-    async def exists(self, job_id: UUID) -> bool:
-        """Check if job exists."""
-        ...
+logger = structlog.get_logger()
 
 
 class RedisJobRepository:
-    """
-    Redis-based job repository.
+    """Repository for batch job persistence in Redis."""
 
-    Stores jobs as JSON with TTL.
-    """
-
-    def __init__(self, redis: Redis, prefix: str = "pdfsvc", ttl_seconds: int = 3600):
-        self._redis = redis
-        self._prefix = prefix
-        self._ttl = ttl_seconds
-
-    def _key(self, job_id: UUID) -> str:
-        """Generate Redis key for job."""
-        return f"{self._prefix}:job:{job_id}"
-
-    async def save(self, job: Job) -> None:
+    def __init__(
+        self,
+        redis_client: Redis,
+        settings: Settings,
+    ) -> None:
         """
-        Save job to Redis with TTL.
+        Initialize job repository.
 
         Args:
-            job: Job instance to save
+            redis_client: Connected Redis client
+            settings: Application settings
         """
-        key = self._key(job.id)
+        self.redis = redis_client
+        self.settings = settings
+        self.key_prefix = settings.redis.key_prefix
+        self.ttl = settings.redis.job_ttl_seconds
+
+    def _job_key(self, job_id: UUID) -> str:
+        """Generate Redis key for a job."""
+        return f"{self.key_prefix}:job:{job_id}"
+
+    async def save(self, job: BatchJob) -> None:
+        """
+        Save a batch job to Redis.
+
+        Args:
+            job: BatchJob to save
+        """
+        key = self._job_key(job.job_id)
         data = orjson.dumps(job.model_dump(mode="json"))
 
-        await self._redis.setex(key, self._ttl, data)
-        logger.debug("job_saved", job_id=str(job.id), status=job.status)
+        await self.redis.setex(
+            key,
+            self.ttl,
+            data,
+        )
 
-    async def get(self, job_id: UUID) -> Optional[Job]:
+        logger.debug(
+            "job_saved",
+            job_id=str(job.job_id),
+            status=job.status.value,
+            key=key,
+        )
+
+    async def get(self, job_id: UUID) -> BatchJob | None:
         """
-        Get job by ID from Redis.
+        Get a batch job by ID.
 
         Args:
             job_id: Job UUID
 
         Returns:
-            Job instance or None if not found
+            BatchJob if found, None otherwise
         """
-        key = self._key(job_id)
-        data = await self._redis.get(key)
+        key = self._job_key(job_id)
+        data = await self.redis.get(key)
 
-        if not data:
-            logger.debug("job_not_found", job_id=str(job_id))
+        if data is None:
             return None
 
         job_dict = orjson.loads(data)
-        job = Job.model_validate(job_dict)
-        logger.debug("job_retrieved", job_id=str(job_id), status=job.status)
-        return job
+        return BatchJob.model_validate(job_dict)
 
     async def delete(self, job_id: UUID) -> bool:
         """
-        Delete job from Redis.
+        Delete a job by ID.
 
         Args:
             job_id: Job UUID
@@ -98,15 +96,13 @@ class RedisJobRepository:
         Returns:
             True if deleted, False if not found
         """
-        key = self._key(job_id)
-        result = await self._redis.delete(key)
-        deleted = result > 0
-        logger.debug("job_deleted", job_id=str(job_id), success=deleted)
-        return deleted
+        key = self._job_key(job_id)
+        result = await self.redis.delete(key)
+        return result > 0
 
     async def exists(self, job_id: UUID) -> bool:
         """
-        Check if job exists in Redis.
+        Check if a job exists.
 
         Args:
             job_id: Job UUID
@@ -114,64 +110,134 @@ class RedisJobRepository:
         Returns:
             True if exists
         """
-        key = self._key(job_id)
-        return await self._redis.exists(key) > 0
+        key = self._job_key(job_id)
+        return await self.redis.exists(key) > 0
 
-    async def update_status(self, job_id: UUID, status: str, **kwargs) -> Optional[Job]:
+    async def update_status(
+        self,
+        job_id: UUID,
+        status: JobStatus,
+    ) -> bool:
         """
-        Update job status and save.
+        Update job status.
 
         Args:
             job_id: Job UUID
             status: New status
-            **kwargs: Additional fields to update
 
         Returns:
-            Updated job or None if not found
+            True if updated, False if job not found
         """
         job = await self.get(job_id)
-        if not job:
-            return None
+        if job is None:
+            return False
 
         job.status = status
-        for key, value in kwargs.items():
-            if hasattr(job, key):
-                setattr(job, key, value)
-
         await self.save(job)
-        return job
+        return True
 
-    async def get_all_by_status(self, status: str) -> list[Job]:
+    async def get_all_by_status(
+        self,
+        status: JobStatus,
+        limit: int = 100,
+    ) -> list[BatchJob]:
         """
         Get all jobs with a specific status.
 
-        Note: This scans keys, use sparingly.
-
         Args:
-            status: Job status to filter by
+            status: Status to filter by
+            limit: Maximum number of jobs to return
 
         Returns:
-            List of jobs with matching status
+            List of BatchJobs matching status
         """
-        pattern = f"{self._prefix}:job:*"
+        pattern = f"{self.key_prefix}:job:*"
         jobs = []
 
-        async for key in self._redis.scan_iter(match=pattern):
-            data = await self._redis.get(key)
+        async for key in self.redis.scan_iter(pattern, count=limit):
+            data = await self.redis.get(key)
             if data:
                 job_dict = orjson.loads(data)
-                if job_dict.get("status") == status:
-                    jobs.append(Job.model_validate(job_dict))
+                job = BatchJob.model_validate(job_dict)
+                if job.status == status:
+                    jobs.append(job)
+                    if len(jobs) >= limit:
+                        break
 
-        logger.debug("jobs_by_status", status=status, count=len(jobs))
         return jobs
 
+    async def get_pending_jobs(self, limit: int = 100) -> list[BatchJob]:
+        """
+        Get all pending jobs.
 
-def create_job_repository(redis: Redis) -> RedisJobRepository:
-    """Factory function to create job repository."""
-    settings = get_settings()
-    return RedisJobRepository(
-        redis=redis,
-        prefix=settings.redis.key_prefix,
-        ttl_seconds=settings.redis.job_ttl_seconds,
-    )
+        Args:
+            limit: Maximum number of jobs to return
+
+        Returns:
+            List of pending BatchJobs
+        """
+        return await self.get_all_by_status(JobStatus.PENDING, limit)
+
+    async def get_processing_jobs(self, limit: int = 100) -> list[BatchJob]:
+        """
+        Get all processing jobs.
+
+        Args:
+            limit: Maximum number of jobs to return
+
+        Returns:
+            List of processing BatchJobs
+        """
+        return await self.get_all_by_status(JobStatus.PROCESSING, limit)
+
+    async def enqueue_job(self, job: BatchJob) -> None:
+        """
+        Save job and add to processing queue.
+
+        Args:
+            job: BatchJob to enqueue
+        """
+        # Save job
+        await self.save(job)
+
+        # Add to queue
+        queue_key = f"{self.key_prefix}:{self.settings.redis.queue_pdf_jobs}"
+        await self.redis.rpush(queue_key, str(job.job_id))
+
+        logger.info(
+            "job_enqueued",
+            job_id=str(job.job_id),
+            queue=queue_key,
+            total_items=job.total_items,
+        )
+
+    async def dequeue_job(self, timeout: float = 5.0) -> BatchJob | None:
+        """
+        Get next job from processing queue.
+
+        Args:
+            timeout: Timeout in seconds for blocking pop
+
+        Returns:
+            BatchJob if available, None otherwise
+        """
+        queue_key = f"{self.key_prefix}:{self.settings.redis.queue_pdf_jobs}"
+
+        result = await self.redis.blpop(queue_key, timeout=timeout)
+        if result is None:
+            return None
+
+        _, job_id_bytes = result
+        job_id = UUID(job_id_bytes.decode())
+
+        return await self.get(job_id)
+
+    async def get_queue_length(self) -> int:
+        """
+        Get number of jobs in the queue.
+
+        Returns:
+            Queue length
+        """
+        queue_key = f"{self.key_prefix}:{self.settings.redis.queue_pdf_jobs}"
+        return await self.redis.llen(queue_key)
