@@ -1,67 +1,80 @@
-"""Models module."""
+"""Models package."""
 
 from pdf_svc.models.events import (
     BaseEvent,
     FileDownloadCompleted,
-    FileDownloadCompletedPayload,
     FileDownloadFailed,
-    FileDownloadFailedPayload,
+    FileDownloadPayload,
     FileDownloadRequest,
     FileDownloadRequestPayload,
     FileUploadCompleted,
-    FileUploadCompletedPayload,
     FileUploadFailed,
-    FileUploadFailedPayload,
+    FileUploadPayload,
     FileUploadRequest,
     FileUploadRequestPayload,
-    PdfBatchProcessRequest,
-    PdfItem,
-    PdfJobStatusPayload,
-    PdfJobStatusRequest,
-    PdfJobStatusResponse,
-    PdfProcessCompleted,
-    PdfProcessCompletedPayload,
-    PdfProcessFailed,
-    PdfProcessFailedPayload,
-    PdfProcessRequest,
-    PdfProcessRequestPayload,
-    QrConfigItem,
-    QrPdfConfigItem,
+    JobStatusRequest,
+    JobStatusRequestPayload,
+    JobStatusResponse,
+    JobStatusResponsePayload,
+    PdfBatchCompleted,
+    PdfBatchFailed,
+    PdfBatchRequest,
+    PdfBatchRequestPayload,
+    PdfBatchResultPayload,
+    PdfItemCompleted,
+    PdfItemCompletedPayload,
+    PdfItemFailed,
+    PdfItemFailedPayload,
+    PdfItemRequest,
+    PdfItemResult,
 )
-from pdf_svc.models.job import Job, JobError, JobResult, JobStage, JobStatus
+from pdf_svc.models.job import (
+    BatchItem,
+    BatchJob,
+    ItemData,
+    ItemError,
+    ItemStatus,
+    JobStatus,
+)
 
 __all__ = [
     # Job models
-    "Job",
+    "BatchJob",
+    "BatchItem",
+    "ItemData",
+    "ItemError",
+    "ItemStatus",
     "JobStatus",
-    "JobStage",
-    "JobError",
-    "JobResult",
-    # Event models
+    # Base event
     "BaseEvent",
-    "FileDownloadRequest",
-    "FileDownloadRequestPayload",
+    # File-svc events (inbound)
+    "FileDownloadPayload",
     "FileDownloadCompleted",
-    "FileDownloadCompletedPayload",
     "FileDownloadFailed",
-    "FileDownloadFailedPayload",
-    "FileUploadRequest",
-    "FileUploadRequestPayload",
+    "FileUploadPayload",
     "FileUploadCompleted",
-    "FileUploadCompletedPayload",
     "FileUploadFailed",
-    "FileUploadFailedPayload",
-    "PdfProcessRequest",
-    "PdfProcessRequestPayload",
-    "PdfBatchProcessRequest",
-    "PdfProcessCompleted",
-    "PdfProcessCompletedPayload",
-    "PdfProcessFailed",
-    "PdfProcessFailedPayload",
-    "PdfJobStatusRequest",
-    "PdfJobStatusResponse",
-    "PdfJobStatusPayload",
-    "PdfItem",
-    "QrConfigItem",
-    "QrPdfConfigItem",
+    # File-svc events (outbound)
+    "FileDownloadRequestPayload",
+    "FileDownloadRequest",
+    "FileUploadRequestPayload",
+    "FileUploadRequest",
+    # PDF-svc events (inbound)
+    "PdfItemRequest",
+    "PdfBatchRequestPayload",
+    "PdfBatchRequest",
+    # PDF-svc events (outbound)
+    "PdfItemResult",
+    "PdfBatchResultPayload",
+    "PdfBatchCompleted",
+    "PdfBatchFailed",
+    "PdfItemCompletedPayload",
+    "PdfItemCompleted",
+    "PdfItemFailedPayload",
+    "PdfItemFailed",
+    # Job status events
+    "JobStatusRequestPayload",
+    "JobStatusRequest",
+    "JobStatusResponsePayload",
+    "JobStatusResponse",
 ]
