@@ -1,5 +1,5 @@
 """Shared utilities module."""
 
-from pdf_svc.shared.logger import configure_logging, get_logger, logger
+from pdf_svc.shared.logger import configure_logging, get_logger, logger, setup_logging
 
-__all__ = ["configure_logging", "get_logger", "logger"]
+__all__ = ["configure_logging", "get_logger", "logger", "setup_logging"]
